feature: add matching option
Add an optional `matching` option to find the latest commit matching
a specific commit message.

This is beneficial to users who have in-repo content (eg. in YAML files)
and display the date of last modification with `gatsby-transformer-gitinfo`.

Editor or developers can tag commits with a specific label
(eg. "non-content") to indicate changes that shouldn't affect the
displayed date, like changes in comments or in the file structure.